Convert BillboardData to be a Resource instead of a Component.
